+++ sparclur/__init__.py
@@ -3,14 +3,18 @@ from ._detect_chaos import DetectChaos
 from ._spotlight import Spotlight
 from ._roll_back import RollBack
 from ._floodlight import FloodLight
+from ._reports import BatchReport, DocumentReport, SparclurReport
 
 __version__ = '2026.9.9'
 
 __all__ = [
     "Astrotruther",
+    "BatchReport",
     "DetectChaos",
+    "DocumentReport",
     "FloodLight",
     "RollBack",
+    "SparclurReport",
     "Spotlight",
     "__version__",
 ]